requirements.txt
# Core Dependencies
neo4j>=5.0.0
networkx>=2.8.0
pandas>=1.5.0
numpy>=1.23.0

# NLP and ML
scikit-learn>=1.0.0
nltk>=3.8.0
spacy>=3.5.0
torch>=2.0.0

# Web and API
flask>=3.0.0
werkzeug>=3.0.0
requests>=2.0.0
flask-cors==3.0.10

# Utilities
python-dotenv>=0.19.0
tqdm>=4.65.0

# Testing and Development
pytest>=7.0.0
pytest-cov>=4.1.0
black>=23.3.0
jupyter>=1.0.0
ipykernel>=6.0.0
sentence-transformers==2.2.2

# Added from the code block
matplotlib==3.4.3
PyYAML==6.0 